requirements.txt
asgiref==3.8.1
babel==2.16.0
Django==5.1.3
django-cors-headers==4.6.0
django-modeltranslation==0.19.11
django-money==3.5.3
djangorestframework==3.15.2
pillow==11.0.0
psycopg2-binary==2.9.10
py-moneyed==3.0
python-dotenv==1.0.1
setuptools==75.6.0
sqlparse==0.5.2
typing_extensions==4.12.2
